Uploading files of the new version 0.2.2

+++ include/io/file_info.hpp
@@ -6,5 +6,5 @@
 struct FileInfo {
     std::string name{};
     FileType type{};
-    uint64_t size{};
+    size_t size{};
 };
